requirements.txt
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.1
psycopg2-binary==2.9.7
pydantic~=2.3.0
pandas~=2.1.1
tqdm~=4.66.1
langdetect~=1.0.9
numpy~=1.25.2
SQLAlchemy==2.0.23
openai==1.8.0
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langchain-openai==0.0.3
fastparquet
pandarallel
umap-learn==0.5.5
